refactor(data-migrations): rename ProtocolClassifier to ProtocolValidator
  Restore the existing terminology that the data-migration doc and design doc
  have used all along. The prior refactor introduced "Classifier" as a new
  name for the per-protocol seam — that term was made up. The doc has always
  called this step "Validator," and the framework's process verb is
  "classification" (which stays).

  Renames:
  - ProtocolClassifier → ProtocolValidator
  - ClassifyBatch → Validate
  - ClassificationInput/Result → ValidationInput/Result
  - Register/Get/Build Classifier(s) → Register/Get/Build Validator(s)
  - IngestServiceConfig.ProtocolClassifiers → ProtocolValidators
  - Files: protocol_classifier.go merged into protocol_validator.go;
    classifier_registry.go → validator_registry.go;
    protocol_classify_dispatch.go → protocol_validation_dispatch.go;
    sep41/classifier.go collapsed into sep41/validator.go

  Kept: DispatchClassification, runClassification — "classification" is the
  framework's process verb, distinct from the per-protocol Validator noun.

  SEP-41 consolidation: the exported Validator struct now satisfies
  services.ProtocolValidator directly. The pure SEP-41 signature check lives
  in matchSEP41Spec(specs) for unit-testing in isolation. Metadata
  enrichment fields and methods moved onto Validator; the wrapper struct
  introduced by the prior refactor is gone.

  Doc: adding-a-protocol.md Step 2 ("Create a Protocol Validator") and
  Step 5 ("Register the Validator and Processor") restored to their
  original titles, with the body updated to describe the new richer
  Validate signature. File checklist points at validator.go.

+++ internal/serve/graphql/resolvers/account_sep41_test.go
@@ -317,7 +317,7 @@ func TestAccountResolver_SEP41AllowancesPaginates(t *testing.T) {
 	require.Len(t, rpage2.Edges, 2)
 
 	// Oversize page is rejected with BAD_USER_INPUT.
-	huge := int32(maxAllowancePageLimit + 1)
+	huge := maxAllowancePageLimit + 1
 	_, err = resolver.Sep41Allowances(ctx, parentAccount, &huge, nil, nil, nil)
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "less than or equal to 100")
